refactor(rtc): consolidate RTT mutation ownership
