+++ packages/shared-server/rallar-system/rtc-topology/inbox/rtc-rtt-app-inbox-handler.ts
@@ -1,10 +1,9 @@
 import type { PSqlTransactionSql } from '../../../postgres/PostgresSqlClient.ts';
-import { readRttMutation, writeRttMutation } from '../../services/rtc-rtt-mutation-service.ts';
-import { computeRttMutation, validateRttMutation } from '../../services/rtc-topology-mutations.ts';
-import {
-  type RtcRttAppInboxResult,
-  toRtcRttAppInboxResult,
-} from '../../services/rtc-rtt-app-inbox-result.ts';
+import { computeRtcRttMutation } from '../mutation/compute-rtc-rtt-mutation.ts';
+import { readRtcRttMutation } from '../mutation/read-rtc-rtt-mutation.ts';
+import { validateRtcRttMutation } from '../mutation/validate-rtc-rtt-mutation.ts';
+import { writeRtcRttMutation } from '../mutation/write-rtc-rtt-mutation.ts';
+import { type RtcRttAppInboxResult, toRtcRttAppInboxResult } from './rtc-rtt-app-inbox-result.ts';
 import type { AppInboxMessageContext } from '../../services/AppInboxService.ts';
 import type { GroupStateService } from '../../group-state/group-state-service-contracts.ts';
 import {
@@ -60,7 +59,7 @@ export class RtcRttAppInboxHandler {
       rtt: authority.command.rtt,
       alSenderId: authority.command.actor.sessionId,
     };
-    const read = await readRttMutation(rtcRttDependencies.repository, stableRequest);
+    const read = await readRtcRttMutation(rtcRttDependencies.repository, stableRequest);
     const attemptCount = context.entry.dequeueAudit.attempts;
     const command = read.receipt
       ? {
@@ -81,8 +80,8 @@ export class RtcRttAppInboxHandler {
       commandHash: authority.command.mutationCommandHash,
       attemptCount,
     };
-    const computed = computeRttMutation({ command, read, facts });
-    validateRttMutation({ command, read, facts, computed });
+    const computed = computeRtcRttMutation({ command, read, facts });
+    validateRtcRttMutation({ command, read, facts, computed });
     const result = await this.commitMutation(context, authority.command, computed, facts);
     if (computed.outcome === 'write') {
       rtcRttDependencies.observeCommitted?.(computed.measurementGuard.value);
@@ -101,15 +100,15 @@ export class RtcRttAppInboxHandler {
   private async commitMutation(
     context: AppInboxMessageContext,
     command: RtcRttAppInboxCommand,
-    computed: ReturnType<typeof computeRttMutation>,
-    facts: Parameters<typeof computeRttMutation>[0]['facts'],
+    computed: ReturnType<typeof computeRtcRttMutation>,
+    facts: Parameters<typeof computeRtcRttMutation>[0]['facts'],
   ): Promise<RtcRttAppInboxResult> {
     return await this.dependencies.writeMutation(context, async (transaction) => {
       if (computed.outcome === 'write') {
         if (facts.requestedAtEpochMs === null || facts.purgeAfterEpochMs === null) {
           throw new TypeError('RTC RTT write lifecycle facts are missing');
         }
-        await writeRttMutation(
+        await writeRtcRttMutation(
           transaction,
           {
             ttlMs: facts.purgeAfterEpochMs - facts.requestedAtEpochMs,
