+++ packages/shared-server/rallar-system/rtc-topology/topic/init-rtc-rtt-topic.ts
@@ -18,7 +18,9 @@ import type { RtcRttRefinementGate } from './rtc-rtt-refinement-gate.ts';
 
 import type { RtcTopologyRuntimeState } from '../../ws-rtc-topology-runtime.ts';
 import type { RtcTopologyWorkPublisher } from '../../services/RtcTopologyOutboxWork.ts';
-import type { GroupTopologyGroupSnapshotReader } from '../../topology/group-topology-management-service.ts';
+// prettier-ignore
+import type { GroupTopologyGroupSnapshotReader }
+  from '../../topology/group-topology-management-service.ts';
 import type { RallarRtcTopologyService } from '../../services/rallar-rtc-topology-service.ts';
 import {
   evaluateRtcRttMeasurement,
